+++ src/Types/isNull.ts
@@ -1,0 +1,11 @@
+/**
+ * Created by Martin Neundorfer on 10.01.2019.
+ * For LABOR.digital
+ */
+/**
+ * Returns true if the given value is NULL, false if not
+ * @param value
+ */
+export function isNull(value):boolean {
+	return value === null
+}